# sem11/openapi.yaml
openapi: 3.0.3
info:
  title: Robot service 3.0
  description: |-
    API сервис управления роботом - пылесосом

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://localhost:9669
...
paths:
...
  /Group:
    put:
      tags:
        - Group
      summary: Обновить группу
      description: Обновить группу
      operationId: updateGroup
      requestBody:
        description: Обновить существующую группу в базе
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
          application/xml:
            schema:
              $ref: '#/components/schemas/Group'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid ID supplied
        '404':
          description: Group not found
        '405':
          description: Validation exception
        
    post:
      tags:
        - Group
      summary: Добавить группу
      description: Добавить группу в базу
      operationId: addGroup
      requestBody:
        description: Создание новой группы в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
          application/xml:
            schema:
              $ref: '#/components/schemas/Group'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Group'
        '405':
          description: Invalid input
  /Group/findById/{groupId}:
    get:
      tags:
        - Group
      summary: Найти группу по ID
      description: вернуть группу
      operationId: getGroupById
      parameters:
        - name: groupId
          in: path
          description: Идентификатор группы
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid ID supplied
        '404':
          description: Group not found
  /Group/deleteById/{groupId}:      
    delete:
      tags:
        - Group
      summary: Удаляет группу
      description: Удаляет группу
      operationId: deleteGroup
      parameters:
        - name: groupId
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: groupId
          in: path
          description: ID группы
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid Group value
...
  /Cleaning:
    put:
      tags:
        - Сleaning
      summary: Обновить уборку
      description: Обновить убоку
      operationId: updateСleaning
      requestBody:
        description: Создание уборки в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cleaning'
          application/xml:
            schema:
              $ref: '#/components/schemas/Cleaning'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Cleaning'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cleaning'  
            application/xml:
              schema:
                $ref: '#/components/schemas/Cleaning'
        '400':
          description: Invalid ID supplied
        '404':
          description: Robot not found
        '405':
          description: Validation exception
    post:
      tags:
        - Сleaning
      summary: Добавить уборку
      description: Добавить уборку по ID
      operationId: addCleaning
      requestBody:
        description: Добавить уборку в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cleaning'
          application/xml:
            schema:
              $ref: '#/components/schemas/Cleaning'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Cleaning'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cleaning' 
            application/xml:
              schema:
                $ref: '#/components/schemas/Cleaning'
        '405':
          description: Invalid input
...
components:
  schemas:
...
    Group:
      required:
        - userId
        - robotId
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        userId:
          type: integer
          format: int64
          example: 10
        robotId:
          type: integer
          format: int64
          example: 10  
      xml:
        name: Group
...
    Cleaning:
      required:
        - dateTime
        - type
        - description
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        dateTime:
          type: string
          example: dd.mm.yyyy-hh:mm:ss
        type:
          type: string
          example: something
        description:
          type: string
          example: something
      xml:
        name: Cleaning
...