# sem11/src/spring-server-generated/src/main/java/io/swagger/api/CleaningApiController.java
package io.swagger.api;

import io.swagger.model.Cleaning;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import javax.servlet.http.HttpServletRequest;
import java.io.IOException;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-08-12T11:28:20.035575131Z[GMT]")
@RestController
public class CleaningApiController implements CleaningApi {

    private static final Logger log = LoggerFactory.getLogger(CleaningApiController.class);

    private final ObjectMapper objectMapper;

    private final HttpServletRequest request;

    @org.springframework.beans.factory.annotation.Autowired
    public CleaningApiController(ObjectMapper objectMapper, HttpServletRequest request) {
        this.objectMapper = objectMapper;
        this.request = request;
    }

    public ResponseEntity<Cleaning> addCleaning(@Parameter(in = ParameterIn.DEFAULT, description = "Добавить уборку в БД", required=true, schema=@Schema()) @Valid @RequestBody Cleaning body) {
        String accept = request.getHeader("Accept");
        if (accept != null && accept.contains("application/json")) {
            try {
                return new ResponseEntity<Cleaning>(objectMapper.readValue("{\n  \"dateTime\" : \"dd.mm.yyyy-hh:mm:ss\",\n  \"description\" : \"something\",\n  \"id\" : 10,\n  \"type\" : \"something\"\n}", Cleaning.class), HttpStatus.NOT_IMPLEMENTED);
            } catch (IOException e) {
                log.error("Couldn't serialize response for content type application/json", e);
                return new ResponseEntity<Cleaning>(HttpStatus.INTERNAL_SERVER_ERROR);
            }
        }

        return new ResponseEntity<Cleaning>(HttpStatus.NOT_IMPLEMENTED);
    }

    public ResponseEntity<Cleaning> updateleaning(@Parameter(in = ParameterIn.DEFAULT, description = "Создание уборки в БД", required=true, schema=@Schema()) @Valid @RequestBody Cleaning body) {
        String accept = request.getHeader("Accept");
        if (accept != null && accept.contains("application/json")) {
            try {
                return new ResponseEntity<Cleaning>(objectMapper.readValue("{\n  \"dateTime\" : \"dd.mm.yyyy-hh:mm:ss\",\n  \"description\" : \"something\",\n  \"id\" : 10,\n  \"type\" : \"something\"\n}", Cleaning.class), HttpStatus.NOT_IMPLEMENTED);
            } catch (IOException e) {
                log.error("Couldn't serialize response for content type application/json", e);
                return new ResponseEntity<Cleaning>(HttpStatus.INTERNAL_SERVER_ERROR);
            }
        }

        return new ResponseEntity<Cleaning>(HttpStatus.NOT_IMPLEMENTED);
    }

}